Refactors user role
