+++ src/qris_dockwidget.py
@@ -44,8 +44,10 @@ from qgis.PyQt.QtWidgets import QAbstractItemView, QFileDialog
 from qgis.PyQt.QtCore import pyqtSignal, Qt, QDate
 from qgis.PyQt.QtGui import QStandardItemModel, QStandardItem, QIcon
 
-from .classes.context_menu import ContextMenu
-from .classes.settings import Settings
+from .QRiS.context_menu import ContextMenu
+from .QRiS.settings import Settings
+from .QRiS.qt_user_role import item_code
+from .QRiS.manage_map import add_to_map
 
 from .ui.elevation_dockwidget import ElevationDockWidget
 from .ui.project_extent_dialog import ProjectExtentDlg
@@ -56,12 +58,6 @@ from .ui.design_dialog import DesignDlg
 
 FORM_CLASS, _ = uic.loadUiType(os.path.join(
     os.path.dirname(__file__), 'qris_dockwidget.ui'))
-
-item_code = {'item_type': Qt.UserRole + 1,
-             # INSTANCE - Stores the class instance, the whole thing.
-             'INSTANCE': Qt.UserRole + 2,
-             # item_layer usually refers to display within the QGIS layer tree often for groups
-             'feature_id': Qt.UserRole + 3}
 
 
 class QRiSDockWidget(QtWidgets.QDockWidget, FORM_CLASS):
@@ -95,7 +91,6 @@ class QRiSDockWidget(QtWidgets.QDockWidget, FORM_CLASS):
 
     def build_tree_view(self, qris_project, new_item=None):
         """Builds items in the tree view based on dictionary values that are part of the project"""
-        # TODO resolve this naming - it is stupid and inconsistent throughout
         self.qris_project = qris_project
 
         self.model.clear()
@@ -213,7 +208,12 @@ class QRiSDockWidget(QtWidgets.QDockWidget, FORM_CLASS):
         if new_item is not None:
             selected_item = self._find_item_in_model(new_item)
             if selected_item is not None:
-                self.add_to_map(selected_item)
+                add_to_map(self.qris_project, self.model, selected_item)
+
+    def _find_item_in_model(self, name):
+        """Looks in the tree for an item name passed from the dataChange method."""
+        selected_item = self.model.findItems(name, Qt.MatchRecursive)[0]
+        return selected_item
 
     def closeEvent(self, event):
         self.qris_project = None
@@ -232,33 +232,22 @@ class QRiSDockWidget(QtWidgets.QDockWidget, FORM_CLASS):
         if not idx.isValid():
             return
 
-        item = self.model.itemFromIndex(indexes[0])
-        # project_tree_data = item.data(Qt.UserRole)  # ProjectTreeData object
-        # data = project_tree_data.data  # Could be a QRaveBaseMap, a QRaveMapLayer or just some random data
-        # connect signals to treeView menu items
-        item_type = item.data(item_code['item_type'])
+        model_item = self.model.itemFromIndex(indexes[0])
+        item_type = model_item.data(item_code['item_type'])
+
         if item_type == 'project_root':
             self.menu.addAction('EXPAND_ALL', lambda: self.expand_all())
         elif item_type == "project_extent_folder":
             self.menu.addAction('ADD_PROJECT_EXTENT_LAYER', lambda: self.import_project_extent_layer())
             self.menu.addAction('CREATE_BLANK_PROJECT_EXTENT_LAYER', lambda: self.create_blank_project_extent())
-        elif item_type == "project_layers_folder":
-            self.menu.addAction('ADD_PROJECT_LAYER', lambda: self.import_project_layer())
         elif item_type in ['project_extent', 'Project_Extent']:
-            self.menu.addAction('UPDATE_PROJECT_EXTENT', lambda: self.update_project_extent(item))
-            self.menu.addAction('DELETE_PROJECT_EXTENT', lambda: self.delete_project_extent(item))
-            self.menu.addAction('ADD_TO_MAP', lambda: self.add_to_map(item))
-        elif item_type == "DetrendedRastersFolder":
-            self.menu.addAction('ADD_DETRENDED_RASTER', lambda: self.add_detrended_raster())
-        elif item_type == "DetrendedRaster":
-            self.menu.addAction('EXPLORE_ELEVATIONS', lambda: self.explore_elevations(item))
-            self.menu.addAction('ADD_TO_MAP', lambda: self.add_to_map(item))
-        elif item_type in ["DetrendedRasterSurface", "dam_assessment", "design"]:
-            self.menu.addAction('ADD_TO_MAP', lambda: self.add_to_map(item))
-        elif item_type == "assessments_folder":
-            self.menu.addAction('ADD_ASSESSMENT', lambda: self.add_assessment())
+            self.menu.addAction('UPDATE_PROJECT_EXTENT', lambda: self.update_project_extent(model_item))
+            self.menu.addAction('DELETE_PROJECT_EXTENT', lambda: self.delete_project_extent(model_item))
+            self.menu.addAction('ADD_TO_MAP', lambda: add_to_map(self.qris_project, self.model, model_item))
         elif item_type == "designs_folder":
             self.menu.addAction('ADD_DESIGN', lambda: self.add_design())
+        elif item_type in ["design"]:
+            self.menu.addAction('ADD_TO_MAP', lambda: add_to_map(self.qris_project, self.model, model_item))
         else:
             self.menu.clear()
         self.menu.exec_(self.treeView.viewport().mapToGlobal(position))
@@ -355,90 +344,6 @@ class QRiSDockWidget(QtWidgets.QDockWidget, FORM_CLASS):
         self.settings.iface.addDockWidget(Qt.LeftDockWidgetArea, self.elevation_widget)
         self.elevation_widget.dataChange.connect(self.build_tree_view)
         self.elevation_widget.show()
-
-    def add_to_map(self, selected_item):
-        """Adds selected items from the QRiS tree to the QGIS layer tree and also to the map"""
-        node = QgsProject.instance().layerTreeRoot()
-        parents = self._get_parents(selected_item)
-        parents.append(selected_item)
-        # for each node in parent
-        for item in parents:
-            # TODO I don't think we will need item_layer. just item_type
-            item_type = item.data(item_code['item_type'])
-            # Check the item_type and launch specific add to map function
-            if item_type == 'project_extent':
-                self.add_project_extent_to_map(item, node)
-            elif item_type == 'design':
-                self.add_design_to_map(item, node)
-            elif item_type == 'dam_assessment':
-                self.add_assessment_to_map(item, node)
-            # Check if it is a node that is already there
-            elif any([c.name() == item.text() for c in node.children()]):
-                # if is there set it to the next active node and search the next item
-                node = next(n for n in node.children() if n.name() == item.text())
-            else:
-                # if not add the node as a group
-                new_node = node.addGroup(item.text())
-                # and set it as next active search node
-                node = new_node
-
-    def _get_parents(self, start_item: QStandardItem):
-        """This gets all the parent folder items of a selected item in the QRiS tree and passess them back to the add_to_map function"""
-        parents = []
-        placeholder = start_item.parent()
-        while placeholder is not None and placeholder != self.model.invisibleRootItem():
-            parents.append(placeholder)
-            placeholder = placeholder.parent()
-        parents.reverse()
-        return parents
-
-    def _find_item_in_model(self, name):
-        """Looks in the tree for an item name passed from the dataChange method."""
-        selected_item = self.model.findItems(name, Qt.MatchRecursive)[0]
-        return selected_item
-
-    def add_project_extent_to_map(self, item, node):
-        """adds a project extent to the project extent node"""
-        # check if it's already a child of the parent node and add it if not
-        if not any([c.name() == item.text() for c in node.children()]):
-            # if not just crate the path and then add it
-            full_path = item.data(item_code['INSTANCE']).full_path(self.qris_project.project_path)
-            layer = QgsVectorLayer(full_path, item.text(), 'ogr')
-            QgsProject.instance().addMapLayer(layer, False)
-            # TODO add symbology
-            node.addLayer(layer)
-
-    def add_assessment_to_map(self, item, node):
-        """adds assessments to the map"""
-        assessment_id = item.data(item_code['feature_id'])
-        layer = QgsVectorLayer(self.qris_project.assessments_path + "|layername=dams", "Dams-" + item.text(), "ogr")
-        layer.setSubsetString("assessment_id = " + str(assessment_id))
-        QgsExpressionContextUtils.setLayerVariable(layer, 'parent_id', assessment_id)
-        symbology_path = os.path.join(os.path.dirname(__file__), 'symbology', 'assessments_dams.qml')
-        layer.loadNamedStyle(symbology_path)
-        QgsProject.instance().addMapLayer(layer, False)
-        node.addLayer(layer)
-
-    def add_design_to_map(self, item, node):
-        # TODO be sure to add the appropriate
-        # TODO be sure that you are not adding things twice. i.e., check if each layer exists before adding.
-        """adds designs to the map"""
-        design_id = item.data(item_code['feature_id'])
-        design_layer = QgsVectorLayer(self.qris_project.designs_path + "|layername=designs", "designs", "ogr")
-        structure_layer = QgsVectorLayer(self.qris_project.designs_path + "|layername=structures", "Structures-" + str(design_id), "ogr")
-        complex_layer = QgsVectorLayer(self.qris_project.designs_path + "|layername=complexes", "Complexes-" + str(design_id), "ogr")
-        structure_layer.setSubsetString("design_id = " + str(design_id))
-        complex_layer.setSubsetString("design_id = " + str(design_id))
-        structure_symbology_path = os.path.join(os.path.dirname(__file__), 'symbology', 'designs_structures.qml')
-        complex_symbology_path = os.path.join(os.path.dirname(__file__), 'symbology', 'designs_complexes.qml')
-        structure_layer.loadNamedStyle(structure_symbology_path)
-        complex_layer.loadNamedStyle(complex_symbology_path)
-        QgsExpressionContextUtils.setLayerVariable(structure_layer, 'parent_id', design_id)
-        QgsExpressionContextUtils.setLayerVariable(complex_layer, 'parent_id', design_id)
-        QgsProject.instance().addMapLayer(structure_layer, False)
-        QgsProject.instance().addMapLayer(complex_layer, False)
-        node.addLayer(structure_layer)
-        node.addLayer(complex_layer)
 
     def expand_all(self):
         self.treeView.expand_all()
